[feature] Tela de cadastrar produto em loja concluida, falta mexer nos botões de adicionar e remover loja

+++ src/main/java/com/netculture/netculture/controllers/ControllerProduto.java
@@ -2,23 +2,34 @@ package com.netculture.netculture.controllers;
 
 import com.netculture.netculture.models.Produto;
 import com.netculture.netculture.models.Vendedor;
+import com.netculture.netculture.repositories.RepositoryLoja;
 import com.netculture.netculture.repositories.RepositoryProduto;
 import jakarta.servlet.http.HttpSession;
 import org.bson.types.ObjectId;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+
 
 @Controller
 @RequestMapping("/produto")
 public class ControllerProduto {
-    @Autowired
-    private RepositoryProduto repositoryProduto;
-    @Autowired
-    private HttpSession session;
+   
+    private final RepositoryProduto repositoryProduto;
+    
+    private final HttpSession session;
+
+    private final RepositoryLoja repositoryLoja;
+
+    public ControllerProduto(RepositoryProduto repositoryProduto, HttpSession session, RepositoryLoja repositoryLoja) {
+        this.repositoryProduto = repositoryProduto;
+        this.session = session;
+        this.repositoryLoja = repositoryLoja;
+    }
 
     @GetMapping("/create/view")
     public String createProdutoView(Model m) {
@@ -41,7 +52,10 @@ public class ControllerProduto {
 
         produto.setVendedorId(v.getId());
         repositoryProduto.save(produto);
+        m.addAttribute("vLogado", v);
         m.addAttribute("msg", "Produto cadastrado com sucesso!");
+        m.addAttribute("lojas",
+                repositoryLoja.findByVendedorId(v.getId()));
         return "homeVendedor";
     }
 
@@ -66,10 +80,18 @@ public class ControllerProduto {
         return "listarProdutos";
     }
 
-    @DeleteMapping("/delete/{id}")
+    @PostMapping("/delete/{id}")
     public String deletarProduto(@PathVariable ObjectId id, Model m) {
+        Vendedor v = (Vendedor) session.getAttribute("vLogado");
+        if (v == null) {
+            m.addAttribute("msg", "Erro: Vendedor não logado");
+            return "loginVendedor";
+        }
         repositoryProduto.deleteById(id);
         m.addAttribute("msg", "Produto removido com sucesso!");
+        m.addAttribute("vLogado", v);
+        m.addAttribute("lojas",
+                repositoryLoja.findByVendedorId(v.getId()));
         return "homeVendedor";
     }
 }
